added a sections folder and added individual .css file for all the different sections of the page
