+++ src/App.css
@@ -27,14 +27,16 @@
     }
 }
 
-.intro {
-    background-color: blue;
-    margin: 0;
+html,
+body,
+div {
+    /* Height and width fallback for older browsers. */
     height: 100%;
-}
-
-.App div {
-    /* position: absolute; */
-    /* top: 0;
-    left: 0; */
+    width: 100%;
+    /* Set the height to match that of the viewport. */
+    height: 100vh;
+    /* Set the width to match that of the viewport. */
+    width: 100vw;
+    /* Remove any browser-default margins. */
+    margin: 0;
 }
