+++ src/sections/Contact/Contact.css
@@ -1,0 +1,3 @@
+.contact {
+    background: blue;
+}
