+++ src/sections/Contact/Contact.js
@@ -1,0 +1,10 @@
+import React, { Component } from "react";
+import "./Contact.css";
+
+class Contact extends Component {
+  render() {
+    return <div className="contact">contact</div>;
+  }
+}
+
+export default Contact;
