+++ src/sections/Intro/Intro.css
@@ -1,0 +1,3 @@
+.intro {
+    background: blue;
+}
