+++ src/sections/Intro/Intro.js
@@ -1,0 +1,21 @@
+import React, { Component } from "react";
+import "./Intro.css";
+
+class Intro extends Component {
+  render() {
+    return (
+      <div className="intro">
+        Intro
+        {/* <div className="App-header">
+          <img src={logo} className="App-logo" alt="logo" />
+          <h2>Welcome to the Intro Page</h2>
+        </div>
+        <p className="App-intro">
+          To get started, edit <code>src/Intro.js</code> and save to reload.
+        </p> */}
+      </div>
+    );
+  }
+}
+
+export default Intro;
