+++ src/sections/Projects/Projects.css
@@ -1,0 +1,3 @@
+.projects {
+    background: red;
+}
